requirements.txt
Django==1.10.6
Pillow==4.0.0
PyJWT==1.4.2
argparse==1.2.1
django-bootstrap3==8.2.1
django-smart-selects==1.3.4
django-social-share==1.1.1
django-wysiwyg-redactor==0.5.0
lorem-ipsum-generator==0.3
mysqlclient==1.3.10
oauthlib==2.0.1
olefile==0.44
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
transliterate==1.9
wsgiref==0.1.2
